updated with midiFile User model relationship
