+++ middleware/auth.js
@@ -7,14 +7,8 @@ const User = require("../models/User");
 exports.isAuth = asyncHandler(async (req, res, next) => {
     let token;
     const bearer = req.headers.authorization;
-
-    if (
-        bearer &&
-        bearer[0].toLowerCase().startsWith("bearer")
-    ) {
+    if (bearer ) {
         token = req.headers.authorization.split(" ")[1];
-    } else if(req.cookies.token) {
-        token = req.cookies.token
     }
     // Make sure token is sent
     if (!token) {
@@ -22,6 +16,7 @@ exports.isAuth = asyncHandler(async (req, res, next) => {
     }
     try {
         // Verify token
+        console.log(token)
         const decoded = jwt.verify(token, process.env.JWT_SECRET);
         console.log(decoded)
         req.user = await User.findById(decoded.id);
