+++ middleware/error.js
@@ -1,4 +1,4 @@
-const ErrorResponse = require('../Utils/errorResponse');
+const ErrorResponse = require('../utils/errorResponse');
 const errorHandler = (err, req, res, next) => {
 
     let error = {...err};
