+++ routes/midi.js
@@ -1,7 +1,8 @@
 var express = require('express');
 var router = express.Router();
-var midiController = require('../controllers/generateMidi');
+var midiController = require('../controllers/midiFiles');
+const {isAuth}= require('../middleware/auth');
 
-router.post('/generate_drum_rnn', midiController.generateMidiDrums);
+router.post('/generate_drum_rnn', isAuth, midiController.generateDrumRNN);
 
 module.exports = router;
