+++ routes/users.js
@@ -5,14 +5,14 @@ const { isAuth } = require("../middleware/auth");
 var router = express.Router();
 
 
-router.get('/', userController.getUsers)
-
-router.post('/', userController.createUser)
-
-router.get("/:id", userController.getUser)
-
-router.put("/:id", userController.updateUser)
-
-router.delete("/:id", userController.deleteUser)
+// router.get('/', userController.getUsers)
+//
+// router.post('/', userController.createUser)
+//
+// router.get("/:id", userController.getUser)
+//
+// router.put("/:id", userController.updateUser)
+//
+// router.delete("/:id", userController.deleteUser)
 
 module.exports = router;
